conda_requirements.txt
pytorch
torchvision
numpy
pandas
matplotlib
